add big int update field test
